Arrays.clone methods; Func1, Func2, Func3, Func4 versions of Func interface

+++ src/test/java/net/algart/arrays/demo/PArraysSpeed.java
@@ -425,7 +425,7 @@ public class PArraysSpeed {
                         } else {
                             long[] jia = (long[]) ja1Converted;
                             for (int k = 0; k < n; k++) {
-                                sum16 += getLongTest(jia, k);
+                                sum16 += (int) getLongTest(jia, k);
                             }
                         }
                     }
